product-search-python/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.1+cpu
torchvision==0.22.1+cpu
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
pillow
requests
gradio  